requirements.txt
openai>=1.40.0
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
feedparser>=6.0.11
Jinja2>=3.1.4
beautifulsoup4>=4.12.3
Markdown>=3.6
pypdf>=4.3.1
